+++ frontpage-api/src/main/resources/no/ndla/frontpageapi/db/migration/V1__Create_Initial_tables.sql
