+++ frontpage-api/src/main/scala/no/ndla/frontpageapi/DBMigrator.scala
@@ -20,7 +20,7 @@ trait DBMigrator {
       val flyway = Flyway
         .configure()
         .javaMigrations()
-        .locations("frontpageapi/db/migration")
+        .locations("no/ndla/frontpageapi/db/migration")
         .dataSource(dataSource)
         // Seems like flyway uses datasource.getConnection().getScheme() which is null if the scheme does not exist.
         // Therefore we simply override it with dataSource.getScheme.
