+++ frontpage-api/src/main/scala/no/ndla/frontpageapi/db/migration/V2__convert_subjects_to_object.scala
@@ -5,7 +5,7 @@
  * See LICENSE
  */
 
-package frontpageapi.db.migration
+package no.ndla.frontpageapi.db.migration
 
 import io.circe.generic.auto._
 import io.circe.generic.semiauto._
