+++ frontpage-api/src/main/scala/no/ndla/frontpageapi/db/migration/V4__add_language_to_about.scala
@@ -1,10 +1,11 @@
 /*
- * Part of NDLA ndla.
+ * Part of NDLA frontpage-api.
  * Copyright (C) 2018 NDLA
  *
  * See LICENSE
  */
-package frontpageapi.db.migration
+
+package no.ndla.frontpageapi.db.migration
 
 import io.circe.generic.auto._
 import io.circe.generic.semiauto._
@@ -19,7 +20,7 @@ import scalikejdbc._
 
 import scala.util.{Failure, Success}
 
-class V5__add_meta_description extends BaseJavaMigration {
+class V4__add_language_to_about extends BaseJavaMigration {
 
   implicit val decoder: Decoder[V1_DBFrontPageData] = deriveDecoder
   implicit val encoder: Encoder[V1_DBFrontPageData] = deriveEncoder
@@ -39,9 +40,9 @@ class V5__add_meta_description extends BaseJavaMigration {
   }
 
   def convertSubjectpage(subjectPageData: DBSubjectPage): Option[DBSubjectPage] = {
-    parse(subjectPageData.document).flatMap(_.as[V4_SubjectFrontPageData]).toTry match {
+    parse(subjectPageData.document).flatMap(_.as[V3_SubjectFrontPageData]).toTry match {
       case Success(value) =>
-        val newSubjectPage = V5_SubjectFrontPageData(
+        val newSubjectPage = V4_SubjectFrontPageData(
           id = value.id,
           name = value.name,
           filters = value.filters,
@@ -49,8 +50,7 @@ class V5__add_meta_description extends BaseJavaMigration {
           twitter = value.twitter,
           facebook = value.facebook,
           bannerImage = value.bannerImage,
-          about = value.about,
-          metaDescription = Seq(),
+          about = value.about.map(toNewAboutSubjectFormat).getOrElse(Seq()),
           topical = value.topical,
           mostRead = value.mostRead,
           editorsChoices = value.editorsChoices,
@@ -60,6 +60,10 @@ class V5__add_meta_description extends BaseJavaMigration {
         Some(DBSubjectPage(subjectPageData.id, newSubjectPage.asJson.noSpacesDropNull))
       case Failure(_) => None
     }
+  }
+
+  private def toNewAboutSubjectFormat(aboutSubject: V2_AboutSubject): Seq[V4_AboutSubject] = {
+    Seq(V4_AboutSubject(aboutSubject.title, aboutSubject.description, "nb", aboutSubject.visualElement))
   }
 
   private def update(subjectPageData: DBSubjectPage)(implicit session: DBSession) = {
@@ -72,8 +76,7 @@ class V5__add_meta_description extends BaseJavaMigration {
   }
 }
 
-case class V5_MetaDescription(metaDescription: String, language: String)
-case class V5_SubjectFrontPageData(
+case class V4_SubjectFrontPageData(
     id: Option[Long],
     name: String,
     filters: Option[List[String]],
@@ -82,10 +85,10 @@ case class V5_SubjectFrontPageData(
     facebook: Option[String],
     bannerImage: V2_BannerImage,
     about: Seq[V4_AboutSubject],
-    metaDescription: Seq[V5_MetaDescription],
     topical: Option[String],
     mostRead: List[String],
     editorsChoices: List[String],
     latestContent: Option[List[String]],
     goTo: List[String]
 )
+case class V4_AboutSubject(title: String, description: String, language: String, visualElement: V2_VisualElement)
